app/server/requirements.txt
flask==3.0.2
flask-cors==3.0.10
flask-jwt-extended==4.6.0
boto3==1.34.66
python-dotenv==1.0.1
requests==2.31.0
bcrypt==4.1.2